ending exercise1 about annotations

+++ src/com/level1/exercise1/entities/Face2FaceWorker.java
@@ -8,7 +8,7 @@ public class Face2FaceWorker extends Worker{
     }
 
     @Override
-    public double calculateSalary(int hours){
+    public double calculateSalary(double hours){
         return super.calculateSalary(hours) + gas;
     }
 }
